Create README.md
🚀 Codédex Projects
The best programming tutorials on the web.
Built by developers, for developers.

What are Projects?
Codédex Projects is a catalog of fun, beginner-friendly, and real-world programming tutorials available to developers all over the world. Our goal is to create a mountain of top-notch coding content on the web, from Python to Web Development, VR/AR to Machine Learning.